question about ResearchOps Community account
The ResearchOps 1Password Teams account is about to expire. As it was created before November 26, 2021 we were needing to renew every 2 years. Our emails to both the opensource@1password.com and the help emails have not been responded to and the account says that the trial ends in 3-days. Will you help us figure this out? Many thanks!